+++ packages/ui/src/lib/index.ts
@@ -1,4 +1,5 @@
 export * from "./animations";
 export * from "./haptics";
 export * from "./i18n";
+export * from "./stateSurface";
 export * from "./themedStyles";